pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conductor-evals"
version = "0.1.0"
description = "Run AI evals across multiple models in parallel using Conductor"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Conductor OSS" },
]
keywords = ["ai", "evals", "llm", "conductor", "evaluation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "conductor-python>=1.0.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "requests>=2.28.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://github.com/conductor-sdk/conductor-evals"
Repository = "https://github.com/conductor-sdk/conductor-evals"
Issues = "https://github.com/conductor-sdk/conductor-evals/issues"

[project.scripts]
conductor-eval = "scripts.cli:main"

[tool.setuptools.packages.find]
include = ["scripts*", "workers*", "providers*", "config*", "evals*", "workflows*", "tasks*", "ui*"]

[tool.setuptools.package-data]
config = ["model-presets.json", "*.example.json"]
evals = ["**/*.json"]
workflows = ["*.json"]
tasks = ["*.json"]
ui = ["dist/client/**/*", "dist/client/*"]
